ai-agent/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
psycopg2-binary
pydantic
langgraph
langchain-openai
langchain-core
docker
pygithub
datasets
transformers
unsloth